backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.1.0
langgraph==0.0.26
langsmith==0.0.77
langchain-community==0.0.10
ollama==0.1.7
sqlalchemy==2.0.23
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
websockets==12.0
httpx==0.25.2
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pandas==2.1.4
numpy==1.24.3
beautifulsoup4==4.12.2
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1